add: runlint.cmd, moved: /model to /lib, refactor: layout.tsx
