+++ app/components/mode-toggle.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import { ChevronDownIcon, MoonIcon, SunIcon } from "@radix-ui/react-icons";
 import { useTheme } from "next-themes";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/app/components/ui/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
