+++ app/layout.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { ThemeProvider } from "@/app/components/theme-provider";
 import { cn } from "@/lib/utils";
 import type { Metadata } from "next";
