+++ lib/model/ai.config.ts
