+++ lib/model/ai.types.ts
