+++ package.json
@@ -6,42 +6,48 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "lint:fix": "next lint --fix"
   },
   "dependencies": {
-    "@emoji-mart/data": "^1.2.1",
+    "@emoji-mart/data": "^1.1.2",
     "@emoji-mart/react": "^1.1.1",
-    "@hookform/resolvers": "^3.9.0",
-    "@nanostores/react": "github:ai/react",
-    "@radix-ui/react-avatar": "^1.1.0",
-    "@radix-ui/react-dialog": "^1.1.1",
+    "@hookform/resolvers": "^3.3.4",
+    "@langchain/community": "^0.0.26",
+    "@langchain/core": "^0.1.35",
+    "@mlc-ai/web-llm": "^0.2.24",
+    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.1.1",
-    "@radix-ui/react-scroll-area": "^1.1.0",
-    "@radix-ui/react-select": "^2.1.1",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-tooltip": "^1.1.2",
-    "ai": "^3.3.5",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-popover": "^1.0.7",
+    "@radix-ui/react-scroll-area": "^1.0.5",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@types/react-dom": "^18.3.0",
+    "@types/uuid": "^10.0.0",
+    "ai": "^2.2.34",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "framer-motion": "^11.3.24",
-    "lucide-react": "^0.427.0",
-    "marked": "^14.0.0",
+    "clsx": "^2.1.0",
+    "emoji-mart": "^5.5.2",
+    "framer-motion": "^11.0.3",
+    "lucide-react": "^0.322.0",
     "next": "^14.2.5",
-    "next-themes": "^0.3.0",
-    "openai": "^4.55.4",
+    "next-themes": "^0.2.1",
     "react": "^18",
     "react-code-blocks": "^0.1.6",
     "react-dom": "^18",
-    "react-hook-form": "^7.52.2",
-    "react-resizable-panels": "^2.0.23",
+    "react-hook-form": "^7.50.1",
+    "react-resizable-panels": "^2.0.3",
     "react-textarea-autosize": "^8.5.3",
-    "sonner": "^1.5.0",
-    "tailwind-merge": "^2.4.0",
+    "sharp": "^0.33.2",
+    "sonner": "^1.4.0",
+    "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
-    "uuid": "^10.0.0",
-    "zod": "^3.23.8"
+    "uuid": "^9.0.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
     "@types/node": "^22.2.0",
@@ -50,6 +56,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
     "postcss": "^8.4.41",
+    "marked": "^14.0.0",
     "tailwindcss": "^3.4.9"
   }
 }
