+++ runlint.cmd
@@ -1,0 +1,1 @@
+npm run lint:fix && npm run dev
